% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corte.em.R
\name{cut.in}
\alias{cut.in}
\title{cut.in()}
\usage{
\method{cut}{in}(escore, los, death, unit, days, min = 200, exc.ICU = TRUE,
  complete = FALSE, digits = 5)
}
\arguments{
\item{escore}{Acute Physiology Score.}

\item{los}{Observed length of stay.}

\item{death}{Observed death (must be coded as 0 or 1. 1 for death).}

\item{unit}{The ICU where a patient was admitted.}

\item{days}{Vector with days which one wants to average days range in.}

\item{min}{The minimum quantity of patients  in each severity class (default \code{min} = 200)}

\item{exc.ICU}{logical; If TRUE, ICUs without surviving patients are ignored.}

\item{complete}{logical; If TRUE, shows additional information about severity classes.}

\item{digits}{Integer indicating the number of decimal places to be used in the complete output..}
}
\value{
A vector with cut off severity points. Use function \code{\link[base]{cut}} to apply them to the score punctuation and classify your patients data.
}
\description{
To find cut off points to define severity classes to use in function \code{\link{SRU}} and calculate average days to discharged surviving patients.
}
\examples{
data(icu)
days = seq(1,100)
corte = cut.in(icu$Saps3Points,icu$los,icu$UnitDischargeName,icu$Unit,days,exc.ICU=TRUE)
icu$class=cut(icu$Saps3Points,breaks=corte,include.lowest = T)
SRU(prob=icu$Saps3DeathProbabilityStandardEquation,death=icu$UnitDischargeName,unit=icu$Unit,los=icu$los,originals=F,type = 1,plot=F,score=icu$Saps3Points,class=icu$class)
}

