% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{EMS-package}
\alias{EMS-package}
\title{Expected Mean Squares for Mixed Model Anova}
\description{
Compute tables of expected mean squares for mixed model anova.
}
\author{
S. Wotherspoon.
}
