% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.r
\name{ems}
\alias{ems}
\title{Expected Mean Squares for Anova}
\usage{
ems(formula, data, random = NULL, type = "III", restricted = FALSE)
}
\arguments{
\item{formula}{A formula describing the model.}

\item{data}{A data frame in which the variables specified in the
modle formula will be found.}

\item{random}{A character vector listing the predictors to be
considered random.}

\item{type}{The "type" of test.}

\item{restricted}{Should EMS be computed assuming a restricted or
unrestricted model.}
}
\value{
Returns an object of class "ems" with components
\item{\code{call}}{the matched call.}
\item{\code{restricted}}{logical indicating whether the EMS were computed for the restricted or unrestricted model.}
\item{variables}{a logical vector indicating which model variables are random.}
\item{terms}{a logical vector indicating which model terms are random.}
\item{EMS}{a table of the coefficients in the expansions of the EMS.}
}
\description{
Compute expressions for the expeted means squares in a mixed model.
}
\details{
Computes the coefficients in the expressions for the expected mean
squares for a given dataset.

This approach to analysis of variance is not recommended.
}
