% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{miscellaneous}
\alias{f.num}
\alias{miscellaneous}
\title{Miscellaneous functions for data editing}
\usage{
f.num(num.var)
}
\arguments{
\item{num.var}{A character, or factor variable to be formated as numeric.}
}
\description{
Collections of functions for data editing ussually used as lower levels for other functions.

\code{f.num} is a wrapper to format numeric variables that are stored as character or factor, simultaneously it will try to detect comma spearated and replace it by dots before formating the variable as numeric.
}
\examples{

f.num(c("2,4000","10,0000","5.0400"))

}

