% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.r
\name{project.effects}
\alias{project.effects}
\title{Projection onto Model Space}
\usage{
project.effects(formula, data, type = c("I", "II", "III"))
}
\arguments{
\item{formula}{A formula describing the model.}

\item{data}{A data frame in which the variables specified in the
model formula will be found.}

\item{type}{The "type" of test.}
}
\value{
a list with elements
\item{Q}{the projection matrix.}
\item{assign}{a vector assigning the columns of Q to terms in the model.}
}
\description{
Compute the projection onto the model space.
}
