% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.r
\name{qr.complement}
\alias{qr.complement}
\title{Complement of Column Space of a Matrix}
\usage{
\method{qr}{complement}(X)
}
\arguments{
\item{X}{a matrix.}
}
\value{
A projection matrix.
}
\description{
Compute the projection onto the orthogonal complement of the
column space of the matrix X.
}
