% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.r
\name{qr.projection}
\alias{qr.projection}
\title{Column Space of a Matrix}
\usage{
\method{qr}{projection}(X)
}
\arguments{
\item{X}{a matrix.}
}
\value{
A matrix.
}
\description{
Compute the projection onto the column space of the matrix X.
}
