% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{plot.reclass}
\alias{print.reclass}
\alias{reclass}
\title{reclass()}
\usage{
reclass(x, y, same = TRUE, plot = FALSE, digits = 2, compare = c("SRU",
  "SMR", "BOTH"), decreasing = FALSE)

\method{print}{reclass}(x, ...)

\method{plot}{reclass}(x, ..., xlim = range(x$smr_x), ylim = range(x$sru_x),
  xlab = "SMR", ylab = "SRU", points.arg_x = list(pch = 21, col = "white",
  bg = "yellow", cex = 2), points.arg_y = list(pch = 21, col = "white", bg =
  "yellow", cex = 2), med.arg_x = list(col = "dodgerblue4", lwd = 2, lty = 1),
  med.arg_y = list(col = "dodgerblue4", lwd = 2, lty = 1),
  tert.arg_x = list(col = "darkorange2", lty = 2, lwd = 1),
  tert.arg_y = list(col = "darkorange2", lty = 2, lwd = 1),
  changes.arg_x = list(pch = 21, col = "white", bg = "yellow", cex = 2),
  changes.arg_y = list(pch = 21, col = "white", bg = "yellow", cex = 2),
  text.arg_x = list(labels = x$tab$Rank, cex = 0.6),
  text.arg_y = list(labels = x$tab$Rank, cex = 0.6), worse.arg_x = list(pch
  = 21, col = "white", bg = "tomato", cex = 2), worse.arg_y = list(pch = 21,
  col = "white", bg = "tomato", cex = 2), better.arg_x = list(pch = 21, col =
  "white", bg = "mediumseagreen", cex = 2), better.arg_y = list(pch = 21, col
  = "white", bg = "mediumseagreen", cex = 2), auto.legend = TRUE,
  leg.arg = list(x = "bottomleft", bty = "n", xpd = T, inset = -0.35, ncol =
  1, horiz = F, pch = 19, cex = 0.8, pt.cex = 1.5), main.arg_x = list(main =
  "1? Stage"), main.arg_y = list(main = "2? Stage"))
}
\arguments{
\item{x, y}{Objects of class 'SRU'. Be x from the 1st stage and y from the 2nd.}

\item{same}{logical; If TRUE, compare the same units through two diferents times (default).}

\item{plot}{logical; If TRUE, plots a SMR vs. SRU scatter plot with the ICUs which had your classification changed.}

\item{digits}{Integer indicating the number of decimal places to be used in the output.}

\item{compare}{The way one prefer to benchmark the ICUs: by SRU (default), SMR or both of them.}

\item{decreasing}{logical; Should the sort order of ICU's rank be increasing or decreasing?}

\item{x}{For \code{print.reclass} or \code{plot.reclass}, an object of class 'reclass'.}
}
\value{
A table with information about which ICUs changed from a classification to another, their SMR, SRU and rank among other ICUs.
}
\description{
To compare ICUs  with diferents severities classes or throught two diferents times about your efficiency classification.

\code{plot.reclass} Plots a SMR vs. SRU scatter plot with the ICUs which had your classification changed.

\code{print.reclass} Prints a table with information about which ICUs changed from a classification to another.
}
\examples{
data(icu)
#1st quarter
x <- icu[which(format(as.Date(icu$UnitAdmissionDate),"\%m") \%in\% c("01","02","03")),]
#2nd quarter
y <- icu[which(format(as.Date(icu$UnitAdmissionDate),"\%m") \%in\% c("04","05","06")),]
FirstQ <- SRU(prob = x$Saps3DeathProbabilityStandardEquation, death = x$UnitDischargeName, unit = x$Unit, los = x$los, score = x$Saps3Points, originals = T, type = 1, plot = F)
SecondQ <- SRU(prob = y$Saps3DeathProbabilityStandardEquation, death = y$UnitDischargeName, unit = y$Unit, los = y$los, score = y$Saps3Points, originals = T, type = 1, plot = F)
reclass(x = FirstQ, y = SecondQ)
plot(reclass(x = FirstQ, y = SecondQ))
#' @export
}

