% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRU.R, R/cut_in.R
\name{SRU}
\alias{SRU}
\alias{cut_in}
\alias{plot.SRU}
\alias{print.SRU}
\title{Standardized Resource Use (SRU)}
\usage{
SRU(prob, death, unit, los, los.esp, class, score, plot = FALSE, type = 1,
  digits = 2, digits2 = 5, originals = FALSE)

\method{print}{SRU}(x, ...)

\method{plot}{SRU}(x, ..., xlim = range(x$rates[, 2]),
  ylim = range(x$rates[, 1]), xlab = "SMR", ylab = "SRU",
  points.arg = list(pch = 21, col = "white", bg = "cadetblue3", cex = 1.5),
  med.arg = list(col = "dodgerblue4", lwd = 2, lty = 1), tert.arg = list(col
  = "darkorange2", lty = 2, lwd = 1), auto.legend = TRUE, leg.arg = list(x =
  "top", bty = "n", xpd = NA, inset = -0.15, ncol = 2))

cut_in(score, los, death, unit, days, min = 200, exc.ICU = TRUE,
  complete = FALSE, digits = 5)
}
\arguments{
\item{prob}{Death individual predictions (ranging from 0 to 1) in a vector.}

\item{death}{Observed death (Must be coded as 0 or 1. 1 for death and 0 for survivors).}

\item{unit}{ICU where the patient is admitted.}

\item{los}{Observed length of stay.}

\item{los.esp}{Estimated length of stay.}

\item{class}{Patient's severity range.}

\item{score}{Acute Physiology Score.}

\item{plot}{logical; If TRUE plots a SMR versus SRU scatter plot.}

\item{type}{Way to calculate SRU. If 1, does it as the article (default).}

\item{digits, digits2}{Integer indicating the number of decimal places to be used in the output.}

\item{originals}{logical; If TRUE uses the severity classes and average days as the original article.}

\item{x}{For \code{prin t.SRU}, an object of class 'SRU'.}

\item{...}{Arguments to be passed to methods, such as \code{\link[graphics]{graphical parameters}} (see \code{\link[graphics]{par}}).}

\item{xlim, ylim}{Limits of x and y axis for \code{plot.SRU}.}

\item{xlab, ylab}{Labels of x and y axis for \code{plot.SRU}.}

\item{points.arg}{List of arguments passed to \code{\link[graphics]{points}} for plotting points correponding to units' SMR and SRU for \code{plot.SRU}.}

\item{med.arg}{List of arguments passed to \code{\link[graphics]{abline}} for plotting lines corresponding to SRU and SMR medians for \code{plot.SRU}.}

\item{tert.arg}{List of arguments passed to \code{\link[graphics]{abline}} for plotting lines corresponding to SRU and SMR tertiles for \code{plot.SRU}.}

\item{auto.legend}{logical; If TRUE, prints a legend with \code{leg.arg} argumentes for \code{plot.SRU}.}

\item{leg.arg}{List of arguments passed to \code{\link[graphics]{legend}} for plotting legends corresponding to SRU and SMR medians and tertiles for \code{plot.SRU}.}

\item{days}{For \code{cut_in}. Vector with days which one wants to average days range in.}

\item{min}{For \code{cut_in}. The minimum quantity of patients  in each severity class (default \code{min} = 200)}

\item{exc.ICU}{logical; For \code{cut_in}. If TRUE, ICUs without surviving patients are ignored.}

\item{complete}{logical; For \code{cut_in}. If TRUE, shows additional information about severity classes.}
}
\value{
Two tables: one with information about severity classes, and another with information about ICUs classified as Most Efficient or Least Efficient.
\itemize{
\item \code{Sev} Severity class.
\item\code{Total} Total of patients.
\item \code{Surv} Total of survivors.
\item \code{Total.LOS} Total length of stay (days).
\item \code{AvDays} Average days to produce a survivor.
\item \code{N.Unit} Quantity of ICUs.
\item \code{N.Pat} Quantity of patients.
\item\code{SMR} Standardized Mortality Ratio Mean (standard deviation) - see \code{\link{SMR}}.
\item \code{SRU} Standardized Resource Use Mean (standard deviation).
}

Most Efficient ICUs have SRU,SMR < median. Least Efficient ICUs have SRU,SMR > median.

\code{cut_in} returns a vector with cut off severity points. Use function \code{\link[base]{cut}} to apply them to the score punctuation and classify your patients data.
}
\description{
\code{SRU} Calculates the standardized resource use for ICUs (Intensive Care Units) from information regarding individual patients admissions.

 \code{plot.SRU} Plots a SMR versus SRU scatter plot with its medians and tertiles.

 \code{print.SRU} Prints a object of class 'SRU'.

\code{cut_in} is used to find cut off points to define severity classes to use in function \code{SRU} and calculate average days to discharged surviving patients.
}
\examples{

# Loading the dataset
data(icu)

days = seq(1,100)

corte = cut_in(icu$Saps3Points, icu$los, icu$UnitDischargeName, icu$Unit, days, exc.ICU=TRUE)

icu$class=cut(icu$Saps3Points, breaks = corte, include.lowest = TRUE)

# Removing data with inapropriate values
icu <- icu[-which(icu$los < 0 ),]

# Estimating the SRU
x <- SRU(prob = icu$Saps3DeathProbabilityStandardEquation,
death = icu$UnitDischargeName, unit = icu$Unit,
los = icu$los, score = icu$Saps3Points,
originals = TRUE, type = 1, plot = FALSE);x

plot(x)

# To see the units rankings and individual SMR and SRU, ordering by it SRU
x$rates[order(x$rates$sru),]

# SRU with diferent severity classes created by cut_in function
y <- SRU(prob = icu$Saps3DeathProbabilityStandardEquation,
death = icu$UnitDischargeName, unit = icu$Unit,
los = icu$los, score = icu$Saps3Points,
originals = FALSE, type = 1, plot = FALSE, class = icu$class)
y

rm(x, icu, corte)
}
\author{
Lunna Borges and Pedro Brasil
}
\references{
Rothen HU, Stricker K, Einfalt J, Bauer P, Metnitz PGH, Moreno RP, Takala J (2007) Variability in outcome and resource use in intensive care units. Intensive Care Med 33:1329-1336
}
\seealso{
\code{\link{SMR}}, \code{\link{reclass}}
}

