% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.r
\name{term.matrix}
\alias{term.matrix}
\title{Design Matrix for a Model Term}
\usage{
term.matrix(frame, random, restricted)
}
\arguments{
\item{frame}{A model frame with columns corresponding to the
factors in the term.}

\item{random}{A logical vector indicating which factors are
random.}

\item{restricted}{A logical indicating whether to assume a
restricted or unrestricted model.}
}
\value{
A matrix
}
\description{
Construct the submatrix of the design matrix corresponding to a
single term of the model.
}
