% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.r
\name{print.ems}
\alias{print.ems}
\title{Print Method for ems Objects}
\usage{
\method{print}{ems}(obj, ...)
}
\arguments{
\item{obj}{an ems object.}

\item{...}{currently ignored.}
}
\description{
Print method for ems objects
}
