% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR.R
\name{SMR}
\alias{SMR}
\alias{SMR.table}
\alias{forest.SMR}
\title{Standardized Mortality Ratio (SMR)}
\usage{
SMR(obs.var, pred.var, digits = 5, ci.method = c("Hosmer", "Byar"),
  ci.level = 0.95)

SMR.table(data, group.var, obs.var, pred.var, digits = 5, use.label = TRUE,
  var.labels = attr(data, "var.labels")[match(group.var, names(data))],
  ci.method = c("Hosmer", "Byar"), ci.level = 0.95, reorder = c("no",
  "SMR", "lower.Cl", "upper.Cl"), decreasing = FALSE)

forest.SMR(x, mar1 = c(5.1, 1, 4.1, 1), mar.SMR = c(5.1, 7, 4.1, 1),
  overall.arg = list(x = 0.01, font = 2, las = 1, cex = 1, labels =
  var.labels[1], xpd = NA, adj = 0), NOE.args = list(x = c(0.5, 0.675, 0.85),
  font = 2, las = 1, cex = 1, xpd = NA), var.labels.arg = list(x = 0.01, font
  = 2, las = 1, cex = 1, xpd = NA, adj = 0), cat.labels.arg = list(x = 0.1,
  font = 3, las = 1, cex = 0.95, col = "gray30", xpd = NA, adj = 0),
  smr.xlab = "Standardized Mortality Ratio", seg.col = "blue",
  seg.lty = 1, seg.lwd = 1, pch.type = 18, pch.cex = 2, pch.col = 1,
  cex.var = 1, cex.cat = 0.95, cat.col = "gray30", var.col = "black",
  font.var = 2, font.cat = 3, x.var = 0.01, x.cat = 0.1, x.n = 0.5,
  x.observed = 0.675, x.expected = 0.85, adj.var = 0, adj.cat = 0,
  adj.smr = 1, smr.xlim = "auto", smr.pos = 0.06, grid = TRUE,
  digits = 3)
}
\arguments{
\item{obs.var}{Observed death. Accepted values are 0 (absence) or 1 (presence) in a vector. For \code{SMR.table} is must be a character indicating the name of the variable in data.}

\item{pred.var}{Death individual predictions (ranging from 0 to 1) in a vector. For \code{SMR.table} is must be a character indicating the name of the variable in data.}

\item{digits}{Number of digits for rounding the output.}

\item{ci.method}{Method to estimate the confidence interval. "Hosmer" (default) or "Byar" are acceptable values.}

\item{ci.level}{Level of the confidence interval. Default is 0.95.}

\item{data}{A dataset where pred.var, obs.var and group.var are in.}

\item{group.var}{For \code{SMR.table} is must be a vector indicating the name(s) of the variable(s) in data indicating which variables will form the groups across which SMR will be calculated. Must be factor variables.}

\item{use.label}{Logical. Default is TRUE. For \code{SMR.table} this option will replace the variables names by its lables in var.labels argument.}

\item{var.labels}{A character vector with variables labels. The default is to replace the variable name by the label stored at attr(data, "var.labels "). But one may specify labels directly.}

\item{reorder}{Default is "no". Possible values are: "no", "SMR","lower.Cl", and "upper.Cl". Will make the \code{SMR.table} to be ordered within each varibale by its original order, or by SMR order, or by lower.Cl order, or by upper.Cl.}

\item{decreasing}{Logical. Should the order be decreasing or incresing. See \code{\link[base]{order}}}

\item{x}{For the \code{forest.SMR} this is the output of \code{SMR.table}.}

\item{mar1, mar.SMR}{Values to set the margins (mar parameter) of left and right windows. See \code{\link[graphics]{par}}}

\item{overall.arg}{A list of arguments passed to \code{\link[graphics]{text}} for ploting the overall label. Internally  and 'y' coordinate is replaced.}

\item{NOE.args}{A list of arguments passed to \code{\link[graphics]{text}} for ploting the overall N (number of observations), O (observed eaths) and E (expected deaths). Internally 'labels' and 'y' arguments are replaced.}

\item{var.labels.arg}{A list of arguments passed to \code{\link[graphics]{text}} for ploting the variables labels. Internally  and 'y' coordinate is replaced.}

\item{var.labels.arg}{A list of arguments passed to \code{\link[graphics]{text}} for ploting the categories labels. Internally  and 'y' coordinate is replaced.}
}
\value{
If SMR, then:
\itemize{
\item \code{N} Number of subjects analyzed.
\item\code{O} Observed number of deaths.
\item \code{E} Expected number of deaths.
\item \code{SMR} Standardized mortality ratio.
\item \code{lower.Cl} lower confidence limit.
\item \code{upper.Cl} upper confidence limit.
}

If SMR.table, then a data.frame with the the same information as above, and the additional information: "Variables" (variables names), "Levels" (variables levels).

If forest.SMR, then a plot is returned.
}
\description{
Calculates the standardized mortality ratio and its confidence interval. SMR, for a group, is defined as the ratio of the observed deaths in this group and the sum of the predicted individual probabilities of death by any model.

\code{SMR.table} estimate at once the overall SMR and the SMR across several groups, e.g. ICU units or clinical characteristics. The \code{SMR.table} can be ordered by the SMR estimate or its confidence intervals, facilitating the comparinson of the units ranks.

\code{forest.SMR} shows the \code{SMR.table} output as a forest plot. The plot opens two windows and plot at the left side the values from the \code{SMR.table} and at the right side the points and lines graphically representing each SMR and its confidence interval.
}
\examples{

}
\author{
Lunna Borges and Pedro Brasil
}
\references{

}
\seealso{
\link{SRU}
}

