% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRU.R
\name{SRU}
\alias{SRU}
\alias{plot.SRU}
\alias{print.SRU}
\title{Standardized Resource Use (SRU)}
\usage{
SRU(prob, death, unit, los, los.esp, class, score, plot = FALSE, type = 1,
  digits = 2, digits2 = 5, originals = FALSE)

\method{print}{SRU}(x, ...)

\method{plot}{SRU}(x, ..., xlim = range(x$rates[, 2]),
  ylim = range(x$rates[, 1]), xlab = "SMR", ylab = "SRU",
  points.arg = list(pch = 21, col = "white", bg = "cadetblue3"),
  med.arg = list(col = "dodgerblue4", lwd = 2, lty = 1), tert.arg = list(col
  = "darkorange2", lty = 2, lwd = 1), auto.legend = TRUE, leg.arg = list(x =
  "top", bty = "n", xpd = NA, inset = -0.15, ncol = 2))
}
\arguments{
\item{prob}{Death individual predictions (ranging from 0 to 1) in a vector.}

\item{death}{Observed death (Must be coded as 0 or 1. 1 for death).}

\item{unit}{ICU where the patient is admitted.}

\item{los}{Observed length of stay.}

\item{los.esp}{Estimated length of stay.}

\item{class}{Severity range of the patient.}

\item{score}{Acute Physiology Score.}

\item{plot}{logical; If TRUE plots a SMR versus SRU scatter plot.}

\item{type}{Way to calculate SRU. If 1, does it as the article (default).}

\item{digits, digits2}{Integer indicating the number of decimal places to be used in the output.}

\item{originals}{logical; If TRUE uses the severity classes and average days as the original article.}

\item{x}{Object of class 'SRU'.}
}
\value{
Two tables: one with information about severity classes, and another with information about ICUs classified as Most Efficient or Least Efficient.
\itemize{
\item \code{Sev} Severity class.
\item\code{Total} Total of patients.
\item \code{Surv} Total of survivors.
\item \code{Total.LOS} Total length of stay (days).
\item \code{AvDays} Average days to produce a survivor.
\item \code{N.Unit} Quantity if ICUs.
\item \code{N.Pat} Quantity of patients.
\item\code{SMR} Standardized Mortality Ratio Mean (standard deviation) - see \code{\link{SMR}}.
\item \code{SRU} Standardized Resource Use Mean (standard deviation) - see \code{\link{SRU}}.
}

Most Efficient ICUs have SRU,SMR < median. Least Efficient ICUs have SRU,SMR > median.
}
\description{
Calculates the standardized resource use for ICUs with information about its patients.

 \code{plot.SRU} Plots a SMR versus SRU scatter plot with your medians and tertiles.

 \code{print.SRU} Prints a object of class 'SRU'.
}
\examples{
data(ICU)
icu <- icu[-which(icu$los < 0 ),]
SRU(prob = icu$Saps3DeathProbabilityStandardEquation, death = icu$UnitDischargeName, unit = icu$Unit, los = icu$los, score = icu$Saps3Points, originals = T, type = 1, plot = F)
}
\references{
Rothen HU, Stricker K, Einfalt J, Bauer P, Metnitz PGH, Moreno RP, Takala J (2007) Variability in outcome and resource use in intensive care units. Intensive Care Med 33:1329-1336
}

