% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRU.R
\name{SRU}
\alias{SRU}
\alias{SRUcalc}
\alias{cut_in}
\alias{plot.SRU}
\alias{print.SRU}
\title{Standardized Resource Use (SRU)}
\usage{
SRU(prob, death, unit, los, los.esp, class, score, plot = FALSE, type = 1,
  digits = 2, digits2 = 5, originals = FALSE)

\method{print}{SRU}(x, ...)

\method{plot}{SRU}(x, ..., xlim = range(x$rates[, 2]),
  ylim = range(x$rates[, 1]), xlab = "SMR", ylab = "SRU",
  points.arg = list(pch = 21, col = "white", bg = "cadetblue3", cex = 1.5),
  med.arg = list(col = "dodgerblue4", lwd = 2, lty = 1), tert.arg = list(col
  = "darkorange2", lty = 2, lwd = 1), auto.legend = TRUE, leg.arg = list(x =
  "top", bty = "n", xpd = NA, inset = -0.15, ncol = 2))

cut_in(score, los, death, unit, days, min = 200, exc.ICU = TRUE,
  complete = FALSE, digits = 5)

SRUcalc(prob, death, unit, los, score, digits = 2)
}
\arguments{
\item{prob}{Death individual predictions (ranging from 0 to 1) in a vector.}

\item{death}{Observed death. Accepted values are 0 (absence) or 1 (presence) in a vector.}

\item{unit}{A character or factor variable indicating the ICU where the patient is admitted.}

\item{los}{A numeric variable indicating the observed length of stay for that patient.}

\item{los.esp}{Estimated length of stay. This argument is optional and will be required only if type = 2. If the user has an alternative model to estimate the individual LOS,}

\item{class}{Patient's severity range.}

\item{score}{Acute Physiology Score.}

\item{plot}{Logical; If TRUE plots a SMR versus SRU scatter plot.}

\item{type}{Way to calculate SRU. If 1, it does as the original article (default); which is to estimate the ICU's expected LOS by multiplying the overall average days within each severity class by the total number of admissions in the same severity class in that ICU, and summing the expected LOS for each severity class in that ICU. If 2, then the user must provide the expected LOS for each subject (i.e. from a prediction model), and the function will estimate the ICU's expected LOS as the mean of the individual LOS.}

\item{digits, digits2}{Integer indicating the number of decimals to be used in the output.}

\item{originals}{logical; If TRUE it uses the severity classes and average days as the original article and will override the 'class' argument if any. We recommend not to set it TRUE, unless you really know what you are doing.}

\item{x}{For \code{print.SRU} or \code{plot.SRU}, an object of class 'SRU'.}

\item{...}{Arguments to be passed to methods, such as \code{\link[graphics]{graphical parameters}} (see \code{\link[graphics]{par}}).}

\item{xlim, ylim}{Limits of x and y axis for \code{plot.SRU}.}

\item{xlab, ylab}{Labels of x and y axis for \code{plot.SRU}.}

\item{points.arg}{List of arguments passed to \code{\link[graphics]{points}} for plotting points correponding to ICUs' SMR and SRU for \code{plot.SRU}.}

\item{med.arg}{List of arguments passed to \code{\link[graphics]{abline}} for plotting lines corresponding to SRU and SMR medians for \code{plot.SRU}.}

\item{tert.arg}{List of arguments passed to \code{\link[graphics]{abline}} for plotting lines corresponding to SRU and SMR tertiles for \code{plot.SRU}.}

\item{auto.legend}{Logical; If TRUE, prints a legend with \code{leg.arg} arguments for \code{plot.SRU}.}

\item{leg.arg}{List of arguments passed to \code{\link[graphics]{legend}} for plotting legends corresponding to SRU and SMR medians and tertiles for \code{plot.SRU}.}

\item{days}{For \code{cut_in}, this is a vector with days which one wants to average days range in. See example.}

\item{min}{For \code{cut_in}, this is the minimum quantity of patients in each severity class (default \code{min} = 200)}

\item{exc.ICU}{Logical; For \code{cut_in}, if TRUE, ICUs without surviving patients are ignored.}

\item{complete}{Logical; For \code{cut_in}, if TRUE, shows additional information about severity classes.}
}
\value{
Two tables: one with information about severity classes, and another with information about ICUs classified as Most Efficient or Least Efficient.
\itemize{
\item \code{Sev} Severity class.
\item\code{Total} Total of patients.
\item \code{Surv} Total of survivors.
\item \code{Total.LOS} Total length of stay (days).
\item \code{AvDays} Average days to produce a survivor.
\item \code{N.Unit} Quantity of ICUs.
\item \code{N.Pat} Quantity of patients.
\item\code{SMR} Standardized Mortality Ratio Mean (standard deviation) - see \code{\link{SMR}}.
\item \code{SRU} Standardized Resource Use Mean (standard deviation).
}

Most Efficient ICUs have SRU,SMR < median. Least Efficient ICUs have SRU,SMR > median.

\code{cut_in} returns a vector with cut off severity points. Use function \code{\link[base]{cut}} to apply them to the score punctuation and classify your patients data.

\code{SRUcalc} returns a table with:
\itemize{
\item \code{Unit} ICUs names.
\item\code{SMR or SRU} Standardized Rate.
\item \code{N} Number of subjects analyzed.
\item \code{Observed} Observed number of deaths.
\item \code{Expected} Expected number of deaths.
\item \code{LOS_esp} Expected length of stay.
}
}
\description{
\code{SRU} calculates the standardized resource use for ICUs (Intensive Care Units) from information regarding individual patients admissions. The resource use is represented by the patients' length of stay (LOS). Therefore the SRU for each unit is defined as the observed LOS divided by the expected LOS. To estimate the expected LOS for each ICU one must define severity score, here defined by the SAPS 3 score. But in theory it could be any score/probability estimated to predict death in the whole group. The 'plot.SRU' function will return SRU median and a \code{\link{SMR}} median, and classify each unit in the quadrants formed by these two medians: most efficient (ME) is the lower left quadrant and identify ICUs with SRU and SMR below the median;  least efficient (LE) is the upper right quadrant and identifies ICUs which SRU and SMR are above the median; and least achieving (LE) - the lower right wuadrant, and over achieving (OA) - the upper left quadrant.
 \code{plot.SRU} Plots a \code{\link{SMR}} versus 'SRU' scatter plot with its medians and tertiles.

 \code{print.SRU} Prints a object of class 'SRU'.

 \code{cut_in} is used to find limits to define severity classes which are used in \code{SRU} function to calculate average days to discharged surviving patients. Its rationale is to find the severity classes limits that yelds a desired average days of survivors. At some point in time, we made a study to test if different arrangements of the severity classes would yeld different classifications in the efficiency quadrants. Despite this study did not show any difference from the original approach, we left the function in the package. Therefore, any arbitrary severity classes should yeld theh same results.

 \code{SRUcalc} is a simpler function and returns only SRU value for each ICU.
}
\examples{

# Loading the dataset
data(icu)

days <- seq(1,100)

corte <- cut_in(icu$Saps3Points, icu$los, icu$UnitDischargeName, icu$Unit, days, exc.ICU=TRUE)

icu$class <- cut(icu$Saps3Points, breaks = corte, include.lowest = TRUE)

# Removing data with inapropriate values
icu <- icu[-which(icu$los < 0 ),]

# Estimating the SRU
x <- SRU(prob = icu$Saps3DeathProbabilityStandardEquation,
death = icu$UnitDischargeName, unit = icu$Unit,
los = icu$los, score = icu$Saps3Points,
originals = TRUE, type = 1, plot = FALSE)
x
plot(x)

# To see the units rankings and individual SMR and SRU, ordering by it SRU
x$rates[order(x$rates$sru),]

# SRU with diferent severity classes created by cut_in function
y <- SRU(prob = icu$Saps3DeathProbabilityStandardEquation,
death = icu$UnitDischargeName, unit = icu$Unit,
los = icu$los, score = icu$Saps3Points,
originals = FALSE, type = 1, plot = FALSE, class = icu$class)
y

# Using SRUcalc
SRUcalc(prob = icu$Saps3DeathProbabilityStandardEquation, death = icu$UnitDischargeName, unit = icu$Unit, los = icu$los, score = icu$Saps3Points )

rm(x, icu, corte)
}
\author{
Lunna Borges and Pedro Brasil
}
\references{
Rothen HU, Stricker K, Einfalt J, Bauer P, Metnitz PGH, Moreno RP, Takala J (2007) Variability in outcome and resource use in intensive care units. Intensive Care Med 33:1329-1336
}
\seealso{
\code{\link{SMR}}, \code{\link{reclass}}
}

